listener/requirements.txt
telethon==1.36.0
requests==2.31.0
python-dotenv==1.0.1
